src/gateway/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
chardet==4.0.0
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
idna==2.10
psycopg2-binary==2.9.3
pytz==2022.2.1
requests==2.25.1
sqlparse==0.4.3
urllib3==1.26.12
